--- libs/event/src/lib/components/countdown/countdown.component.scss
@@ -1,3 +1,0 @@
-div {
-  margin-block: 40px;
-}
