--- libs/event/src/lib/components/countdown/counter-display/counter-display.component.html
@@ -1,58 +1,0 @@
-<div class="counter-display" fxLayout="row" fxLayoutAlign="space-between center" fxLayoutGap="24px">
-  <!-- Days --> 
-  <span fxLayout="column" fxLayoutAlign="center center"> 
-    <div class="numberContainer" fxLayout fxLayoutAlign="center center">
-      <p 
-        class="mat-title"  
-        [@testAnim2]="newDays">
-          {{ days }}
-      </p> 
-      <p
-      class="mat-title"
-      *ngIf="newDays !== false"
-      (@testAnim.done)="onAnimationDone()"
-      [@testAnim]>
-        {{ newDays }}
-      </p>
-    </div>
-    <p class="mat-caption">days</p>
-  </span>
-  
-  <p class="mat-title">:</p>
-
-  <!-- hours -->
-  <span fxLayout="column" fxLayoutAlign="center center"> 
-    <div class="numberContainer" fxLayout fxLayoutAlign="center center">
-      <p class="mat-title" [@testAnim2]="newHours">
-        {{ hours }}
-      </p> 
-      <p
-      class="mat-title"
-      *ngIf="newHours !== false"
-      (@testAnim.done)="onAnimationDone()"
-      [@testAnim]>
-        {{ newHours }}
-      </p>
-    </div>
-    <p class="mat-caption">hours</p>
-  </span>
-
-  <p class="mat-title">:</p>
-
-  <!-- Minutes -->
-  <span fxLayout="column" fxLayoutAlign="center center"> 
-    <div class="numberContainer" fxLayout fxLayoutAlign="center center">
-      <p class="mat-title" [@testAnim2]="newMinutes">
-        {{ minutes }}
-      </p> 
-      <p 
-      class="mat-title"
-      *ngIf="newMinutes !== false"
-      (@testAnim.done)="onAnimationDone()"
-      [@testAnim]>
-        {{ newMinutes }}
-      </p>
-    </div>
-    <p class="mat-caption">minutes</p>
-  </span>
-</div>
