+++ libs/ui/src/lib/countdown/countdown.component.html
@@ -1,0 +1,62 @@
+<section *ngIf="(timeLeft$ | async) as t" fxLayout="column" fxLayoutAlign="center center" fxLayoutGap="16px">
+  <span class="mat-body-1">Event will start in :</span>
+  <div class="counter-display" fxLayout="row" fxLayoutAlign="space-between center" fxLayoutGap="24px">
+    <!-- Days --> 
+    <span fxLayout="column" fxLayoutAlign="center center"> 
+      <div class="numberContainer" fxLayout fxLayoutAlign="center center">
+        <p 
+          class="mat-title"  
+          [@numberAnim]="newDays">
+            {{ days | number: "2.0"}}
+        </p> 
+        <p
+        class="mat-title"
+        *ngIf="newDays > -1"
+        (@tempNumberAnim.done)="onAnimationDone()"
+        [@tempNumberAnim]>
+          {{ newDays | number: "2.0" }}
+        </p>
+      </div>
+      <p class="mat-caption">days</p>
+    </span>
+    
+    <p class="mat-title">:</p>
+  
+    <!-- hours -->
+    <span fxLayout="column" fxLayoutAlign="center center"> 
+      <div class="numberContainer" fxLayout fxLayoutAlign="center center">
+        <p class="mat-title" [@numberAnim]="newHours">
+          {{ hours | number: "2.0"}}
+        </p> 
+        <p
+        class="mat-title"
+        *ngIf="newHours > -1"
+        (@tempNumberAnim.done)="onAnimationDone()"
+        [@tempNumberAnim]>
+          {{ newHours | number: "2.0" }}
+        </p>
+      </div>
+      <p class="mat-caption">hours</p>
+    </span>
+  
+    <p class="mat-title">:</p>
+  
+    <!-- Minutes -->
+    <span fxLayout="column" fxLayoutAlign="center center"> 
+      <div class="numberContainer" fxLayout fxLayoutAlign="center center">
+        <p class="mat-title" [@numberAnim]="newMinutes">
+          {{ minutes | number: "2.0"}}
+        </p> 
+        <p 
+        class="mat-title"
+        *ngIf="newMinutes > -1"
+        (@tempNumberAnim.done)="onAnimationDone()"
+        [@tempNumberAnim]>
+          {{ newMinutes | number: "2.0" }}
+        </p>
+      </div>
+      <p class="mat-caption">minutes</p>
+    </span>
+  </div>
+
+</section>
