+++ libs/ui/src/lib/countdown/countdown.component.scss
@@ -1,0 +1,27 @@
+section {
+  margin-block: 40px;
+  
+  .counter-display {
+    border: solid 1px var(--primary);
+    border-radius: 4px;
+    padding: 16px 40px;
+  
+    .mat-caption, .mat-title {
+      margin: 0;
+    }
+  
+    .numberContainer {
+      position: relative;
+      height: 35px;
+      width: 40px;
+      overflow: hidden;
+    
+      p {
+        position: absolute;
+        height: 100%;
+      }
+    }
+  }
+}
+
+
