+++ libs/ui/src/lib/countdown/countdown.module.ts
@@ -2,10 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CountdownComponent } from './countdown.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { CounterDisplayComponent } from './counter-display/counter-display.component';
 
 @NgModule({
-  declarations: [CountdownComponent, CounterDisplayComponent],
+  declarations: [CountdownComponent],
   exports: [CountdownComponent],
   imports: [
     CommonModule,
